create a separate hyperlane-testing crate
