+++ dango/testing/src/lib.rs
@@ -3,7 +3,6 @@ mod code;
 pub mod constants;
 mod crypto;
 mod genesis;
-mod hyperlane;
 mod setup;
 
-pub use {account::*, code::*, crypto::*, genesis::*, hyperlane::*, setup::*};
+pub use {account::*, code::*, crypto::*, genesis::*, setup::*};
